settingpage
updating and improving

+++ Modules/sliderbar.java
@@ -5,8 +5,8 @@ import javax.swing.*;
 
 public class sliderbar {
     JSlider slider;
-    public sliderbar(int max, int average, int spacing){
-        slider = new JSlider(JSlider.HORIZONTAL,0,max,average);
+    public sliderbar(int max, Double average, int spacing){
+        slider = new JSlider(JSlider.HORIZONTAL,0,max,average.intValue());
         slider.setSize(300,4);
         slider.setMinorTickSpacing(1);
         slider.setMajorTickSpacing(spacing);
